+++ packages/ccdi-cde/src/v1.rs
@@ -1,5 +1,8 @@
 //! Common data elements that have a major version of one.
 
+pub mod deposition;
 pub mod file;
+pub mod namespace;
+pub mod organization;
 pub mod sample;
 pub mod subject;